\name{automata-package}
\alias{automata-package}
\alias{automata}
\docType{package}
\title{
\packageTitle{automata}
}
\description{
\packageDescription{automata}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{automata}
\packageIndices{automata}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{automata}

Maintainer: \packageMaintainer{automata}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
